// src/Types/user.types.ts
export type UserType = {
  address: {
    city: string;
    country: string;
    state: string;
    streetAddress: string;
    zipCode: string;
  };
  company: {
    name: string;
    suffix: string;
  };
  email: string;
  id: number;
  imageUrl: string;
  ip: string;
  jobArea: string;
  jobDescriptor: string;
  jobType: string;
  lastName: string;
  name: string;
  prefix: string;
  title: string;
};